requirements.txt
# backend
sentinelhub==3.10.2
fastapi==0.111.0
numpy==1.26.4
Pillow==10.3.0
tensorflow_macos==2.16.1; sys_platform == 'darwin'
tensorflow==2.16.1; sys_platform != 'darwin'
keras==3.2.1
uvicorn==0.30.1
python-dotenv==1.0.0

# frontend
streamlit==1.35.0
pydeck==0.9.1
requests==2.32.3
opencv-python-headless==4.10.0.84
pandas==2.2.2
bytesbufio==1.0.3